FastAPIProject/requirement.txt
fastapi
uvicorn
pydantic
python-dotenv
pytest
pytest httpx